python_requirements.txt
# Python dependencies for GAT model integration
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
transformers>=4.20.0

# Optional dependencies for enhanced functionality
opencv-python>=4.5.0
tqdm>=4.62.0
pandas>=1.3.0
